SEQUOIADBMAINSTREAM-948: control records memory usage
git-svn-id: http://192.168.20.11/sequoiadb/trunk@19938 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/import/impRecordQueue.hpp
@@ -100,7 +100,7 @@ namespace import
          return _size;
       }
 
-      inline INT32 bsonSize()
+      inline INT64 bsonSize()
       {
          return _bsonSize;
       }
@@ -164,7 +164,7 @@ namespace import
       bson**   _array;
       INT32    _capacity;
       INT32    _size;
-      INT32    _bsonSize;
+      INT64    _bsonSize;
       BOOLEAN  _finished;
    };
